% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.effectplot.R
\name{gbm.effectplot}
\alias{gbm.effectplot}
\title{GBM plot of main effects with data distribution overlay.}
\usage{
gbm.effectplot(gbm.model, relimp.range, y.label, num.row, num.col)
}
\arguments{
\item{gbm.model}{A gbm model to be plotted.}

\item{relimp.range}{A vector of length 2 giving upper and lower
relative importance cut-offs}

\item{y.label}{name of response variable}

\item{num.row}{Number of rows in output plot}

\item{num.col}{Number of columns in output plot}
}
\value{
plot of main effects of gbm model.
}
\description{
\code{gbm.effectplot} plots marginal effect for important variables.
}
\examples{
\dontrun{
gbm.effectplot(gbm.model=gbm.model,
relimp.range=c(1,100),num.row=4,num.col=4)
}
}

