% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.int2.R
\name{gbm.int2}
\alias{gbm.int2}
\title{GBM function for assessing 2-way input interaction}
\usage{
gbm.int2(gbm.model, relimp.range, n.trees, n.cores = NULL)
}
\arguments{
\item{gbm.model}{A gbm model to be plotted.}

\item{relimp.range}{A vector of length 2 giving upper and lower
relative importance cut-offs}

\item{n.trees}{number of trees for interaction estimation}

\item{n.cores}{number of cores to utilise if NULL then no parallelisation}

\item{plot.it}{to plot or not to plot that is the question}
}
\value{
Matrix of interactions between inputs.
}
\description{
\code{gbm.int2} Calculates all 2-way interaction metrics for inputs with relative importances within a defined range
}
\examples{
\dontrun{
gbm.int2(gbm.model=gbm1,
relimp.range=c(1,100),n.trees=100,n.cores=NULL)
}
}

