% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gbm.plot.NA.R
\name{gbm.plot.NA}
\alias{gbm.plot.NA}
\title{GBM plot of ordered relative influence and missing values.}
\usage{
gbm.plot.NA(gbm.model, relimp.range)
}
\arguments{
\item{gbm.model}{A gbm model to be plotted.}

\item{relimp.range}{A vector of length 2 giving upper and lower
relative importance cut-offs}
}
\value{
plot of relative influence and NAs of gbm model.
}
\description{
\code{gbm.plot.NA} plots relative influence and NA from gbm.
}
\examples{
\dontrun{
gbm.plot.NA(gbm.model=gbm1,df=data1)
}
}

