% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gbm.pred.survfns.R
\name{gbm.pred.survfns}
\alias{gbm.pred.survfns}
\title{Predict survival functions and distributions for gbm cph models}
\usage{
gbm.pred.survfns(gbm.model, max.time, data.in = NULL, ntrees)
}
\arguments{
\item{gbm.model}{A gbm model to extract survival functions and desities from.}

\item{max.time}{A numeric scalar giving the time point
to which estimates should be extended}

\item{data.in}{A dataframe of inputs for prediction if NULL will use the orginal model dataframe}

\item{ntrees}{The number of trees for predictions}
}
\value{
A data.table with hazard function outputs for each row in the prediction data
}
\description{
\code{gbm.pred.survfns} extracts survival model functions and densities and plots if required
}
\examples{
\dontrun{
gbm.survfns(gbm.model=gbm.model,max.time=100,ntrees=100)
}
}

