% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.relimp.R
\name{gbm.relimp}
\alias{gbm.relimp}
\title{GBM plot of ordered relative influence and missing values.}
\usage{
gbm.relimp(gbm.model, relimp.range)
}
\arguments{
\item{gbm.model}{A gbm model to be plotted.}

\item{relimp.range}{A vector of length 2 giving upper and lower
relative importance cut-offs}
}
\value{
plot of relative influence and NAs of gbm model.
}
\description{
\code{gbm.relimp} plots relative importances and number of NAs from gbm.
}
\examples{
\dontrun{
gbm.relimp(gbm.model=gbm1,df=data1)
}
}

