% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gbm.survfns.R
\name{gbm.survfns}
\alias{gbm.survfns}
\title{GBM extract and plot survival functions and distributions}
\usage{
gbm.survfns(gbm.model, t.eval.max, plot.it = FALSE)
}
\arguments{
\item{gbm.model}{A gbm model to extract survival functions and desities from.}

\item{t.eval.max}{A numeric scalar giving the time point
to which estimates and plots should be extended}

\item{plot.it}{plot the functions. Default is FALSE}
}
\value{
plot of survival functions from gbm model
}
\description{
\code{gbm.survfns} extracts survival model functions and densities and plots if required
}
\examples{
\dontrun{
gbm.survfns(gbm.model=gbm.model,t.evel.max=100,plot.it=TRUE)
}
}

